+++ docs/CHANGELOG.md
@@ -15,6 +15,8 @@
 * `layers.LayerNormalization`: 
   * up-scale precision by default if input is a Float16 or BFloat16. Low-precision
     lead to NaNs when reducing values for normalization. Added also a hyperparameter to configure normalization DType.
+* Package `losses`: 
+  * Added experimental `AdaptivePowerLoss`.
 
 # v0.16.1 - 🎄 2024/12/19 🎄 MatMul fixes
 * MatMul fixed for some edge shape configuration and greatly accelerated in some cases.